+++ src/ConfigurationLoader.php
@@ -23,15 +23,16 @@ namespace TechDivision\Import\Cli;
 use Psr\Log\LogLevel;
 use Rhumsaa\Uuid\Uuid;
 use TechDivision\Import\App\Simple;
-use Symfony\Component\Console\Input\InputInterface;
 use TechDivision\Import\Utils\EntityTypeCodes;
 use TechDivision\Import\Cli\Command\InputOptionKeys;
 use TechDivision\Import\Cli\Command\InputArgumentKeys;
+use TechDivision\Import\Cli\Configuration\LibraryLoader;
 use TechDivision\Import\Cli\Utils\MagentoConfigurationKeys;
+use TechDivision\Import\Configuration\Jms\Configuration;
 use TechDivision\Import\Configuration\Jms\Configuration\Database;
 
 /**
- * The configuration factory implementation.
+ * The configuration loader implementation.
  *
  * @author    Tim Wagner <t.wagner@techdivision.com>
  * @copyright 2016 TechDivision GmbH <info@techdivision.com>
@@ -39,20 +40,8 @@ use TechDivision\Import\Configuration\Jms\Configuration\Database;
  * @link      https://github.com/techdivision/import-cli-simple
  * @link      http://www.techdivision.com
  */
-class ConfigurationLoader
+class ConfigurationLoader extends SimpleConfigurationLoader
 {
-
-    /**
-     * The composer name => Magento Edition mappings.
-     *
-     * @var array
-     */
-    protected $editionMappings = array(
-        'magento2ce'                 => 'CE',
-        'project-community-edition'  => 'CE',
-        'magento2ee'                 => 'EE',
-        'project-enterprise-edition' => 'EE'
-    );
 
     /**
      * The array with the default entity type code => import directory mappings.
@@ -63,24 +52,6 @@ class ConfigurationLoader
         EntityTypeCodes::CATALOG_PRODUCT  => 'products',
         EntityTypeCodes::CATALOG_CATEGORY => 'categories',
         EntityTypeCodes::EAV_ATTRIBUTE    => 'attributes'
-    );
-
-    /**
-     * The array with the default entity type => configuration mapping.
-     *
-     * @var array
-     */
-    protected $defaultConfigurations = array(
-        'ce' => array(
-            EntityTypeCodes::EAV_ATTRIBUTE    => 'techdivision/import-attribute',
-            EntityTypeCodes::CATALOG_PRODUCT  => 'techdivision/import-product',
-            EntityTypeCodes::CATALOG_CATEGORY => 'techdivision/import-category'
-        ),
-        'ee' => array(
-            EntityTypeCodes::EAV_ATTRIBUTE    => 'techdivision/import-attribute',
-            EntityTypeCodes::CATALOG_PRODUCT  => 'techdivision/import-product-ee',
-            EntityTypeCodes::CATALOG_CATEGORY => 'techdivision/import-category-ee'
-        )
     );
 
     /**
@@ -120,222 +91,66 @@ class ConfigurationLoader
         )
     );
 
-    protected $configurationClassName;
-
-    /**
-     * The vendor directory used to load the default configuration files from.
-     *
-     * @var string
-     */
-    protected $vendorDir;
-
-    /**
-     * Initializes the configuration loader with the configuration factory class name.
-     *
-     * @param string $vendorDir                 The vendor directory
-     * @param string $configurationFactoryClass The configuration factory class name
-     */
-    public function __construct($vendorDir, $configurationFactoryClass)
-    {
-        $this->vendorDir = $vendorDir;
-        $this->configurationFactoryClass = $configurationFactoryClass;
-    }
-
-    /**
-     * Return's the vendor directory used to load the default configuration files from.
-     *
-     * @return string The vendor directory
-     */
-    public function getVendorDir()
-    {
-        return $this->vendorDir;
-    }
-
-    /**
-     * The configuration factory class name.
-     *
-     * @return string The configuration factory class name
-     */
-    public function getConfigurationFactoryClass()
-    {
-        return $this->configurationFactoryClass;
-    }
-
     /**
      * Factory implementation to create a new initialized configuration instance.
      *
      * If command line options are specified, they will always override the
      * values found in the configuration file.
      *
-     * @param \Symfony\Component\Console\Input\InputInterface $input          The Symfony console input instance
-     * @param string                                          $entityTypeCode The entity type code to use
-     *
      * @return \TechDivision\Import\Cli\Configuration The configuration instance
      * @throws \Exception Is thrown, if the specified configuration file doesn't exist or the mandatory arguments/options to run the requested operation are not available
      */
-    public function load(InputInterface $input, $entityTypeCode)
+    public function load()
     {
 
-        // load the actual vendor directory
-        $vendorDir = $this->getVendorDir();
-
-        // the path of the JMS serializer directory, relative to the vendor directory
-        $jmsDir = DIRECTORY_SEPARATOR . 'jms' . DIRECTORY_SEPARATOR . 'serializer' . DIRECTORY_SEPARATOR . 'src';
-
-        // try to find the path to the JMS Serializer annotations
-        if (!file_exists($annotationDir = $vendorDir . DIRECTORY_SEPARATOR . $jmsDir)) {
-            // stop processing, if the JMS annotations can't be found
-            throw new \Exception(
-                sprintf(
-                    'The jms/serializer libarary can not be found in one of "%s"',
-                    implode(', ', $vendorDir)
-                )
-            );
-        }
-
-        // register the autoloader for the JMS serializer annotations
-        \Doctrine\Common\Annotations\AnnotationRegistry::registerAutoloadNamespace(
-            'JMS\Serializer\Annotation',
-            $annotationDir
-        );
-
-        // load the configuration factory class name
-        $configurationFactoryClass = $this->getConfigurationFactoryClass();
-
-        // query whether or not, a configuration file has been specified
-        if ($configuration = $input->getOption(InputOptionKeys::CONFIGURATION)) {
-            // load the configuration from the file with the given filename
-            /** @var \TechDivision\Import\ConfigurationInterface $instance */
-            $instance = $configurationFactoryClass::factory($configuration);
-
-        } elseif ($magentoEdition = $input->getOption(InputOptionKeys::MAGENTO_EDITION)) {
-            // use the Magento Edition that has been specified as option
-            /** @var \TechDivision\Import\ConfigurationInterface $instance */
-            $instance = $configurationFactoryClass::factory($this->getDefaultConfiguration($magentoEdition, $entityTypeCode));
-
-            // override the Magento Edition
-            $instance->setMagentoEdition($magentoEdition);
-
-        } else {
-            // finally, query whether or not the installation directory is a valid Magento root directory
-            if (!$this->isMagentoRootDir($installationDir = $input->getOption(InputOptionKeys::INSTALLATION_DIR))) {
-                throw new \Exception(
-                    sprintf(
-                        'Directory "%s" specified with option "--installation-dir" is not a valid Magento root directory',
-                        $installationDir
-                    )
-                );
-            }
-
-            // load the composer file from the Magento root directory
-            $composer = json_decode(file_get_contents($composerFile = sprintf('%s/composer.json', $installationDir)), true);
-
-            // try to load and explode the Magento Edition identifier from the Composer name
-            $explodedVersion = explode('/', $composer[MagentoConfigurationKeys::COMPOSER_EDITION_NAME_ATTRIBUTE]);
-
-            // try to locate Magento Edition
-            if (!isset($this->editionMappings[$possibleEdition = end($explodedVersion)])) {
-                throw new \Exception(
-                    sprintf(
-                        '"%s" detected in "%s" is not a valid Magento Edition, please set Magento Edition with the "--magento-edition" option',
-                        $possibleEdition,
-                        $composerFile
-                    )
-                );
-            }
-
-            // if Magento Edition/Version are available, load them
-            $magentoEdition = $this->editionMappings[$possibleEdition];
-
-            // use the Magento Edition that has been detected by the installation directory
-            /** @var \TechDivision\Import\ConfigurationInterface $instance */
-            $instance = $configurationFactoryClass::factory($this->getDefaultConfiguration($magentoEdition, $entityTypeCode));
-
-            // override the Magento Edition, if NOT explicitly specified
-            $instance->setMagentoEdition($magentoEdition);
-        }
+        // load the configuration instance
+        $instance = parent::load();
 
         // query whether or not an operation name has been specified as command line
         // option, if yes override the value from the configuration file
-        if ($operationName = $input->getArgument(InputArgumentKeys::OPERATION_NAME)) {
+        if ($operationName = $this->input->getArgument(InputArgumentKeys::OPERATION_NAME)) {
             $instance->setOperationName($operationName);
-        }
-
-        // query whether or not a system name has been specified as command line
-        // option, if yes override the value from the configuration file
-        if (($input->hasOptionSpecified(InputOptionKeys::SYSTEM_NAME) && $input->getOption(InputOptionKeys::SYSTEM_NAME)) ||
-            $instance->getSystemName() === null
-        ) {
-            $instance->setSystemName($input->getOption(InputOptionKeys::SYSTEM_NAME));
-        }
-
-        // query whether or not a PID filename has been specified as command line
-        // option, if yes override the value from the configuration file
-        if (($input->hasOptionSpecified(InputOptionKeys::PID_FILENAME) && $input->getOption(InputOptionKeys::PID_FILENAME)) ||
-            $instance->getPidFilename() === null
-        ) {
-            $instance->setPidFilename($input->getOption(InputOptionKeys::PID_FILENAME));
-        }
-
-        // query whether or not a Magento installation directory has been specified as command line
-        // option, if yes override the value from the configuration file
-        if (($input->hasOptionSpecified(InputOptionKeys::INSTALLATION_DIR) && $input->getOption(InputOptionKeys::INSTALLATION_DIR)) ||
-            $instance->getInstallationDir() === null
-        ) {
-            $instance->setInstallationDir($input->getOption(InputOptionKeys::INSTALLATION_DIR));
-        }
-
-        // query whether or not a Magento edition has been specified as command line
-        // option, if yes override the value from the configuration file
-        if ($magentoEdition = $input->getOption(InputOptionKeys::MAGENTO_EDITION)) {
-            $instance->setMagentoEdition($magentoEdition);
         }
 
         // query whether or not a Magento version has been specified as command line
         // option, if yes override the value from the configuration file
-        if ($magentoVersion = $input->getOption(InputOptionKeys::MAGENTO_VERSION)) {
+        if ($magentoVersion = $this->input->getOption(InputOptionKeys::MAGENTO_VERSION)) {
             $instance->setMagentoVersion($magentoVersion);
-        }
-
-        // query whether or not a directory for the source files has been specified as command line
-        // option, if yes override the value from the configuration file
-        if ($sourceDir = $input->getOption(InputOptionKeys::SOURCE_DIR)) {
-            $instance->setSourceDir($sourceDir);
         }
 
         // query whether or not a directory containing the imported files has been specified as command line
         // option, if yes override the value from the configuration file
-        if ($targetDir = $input->getOption(InputOptionKeys::TARGET_DIR)) {
+        if ($targetDir = $this->input->getOption(InputOptionKeys::TARGET_DIR)) {
             $instance->setTargetDir($targetDir);
         }
 
         // query whether or not a directory containing the archived imported files has been specified as command line
         // option, if yes override the value from the configuration file
-        if ($archiveDir = $input->getOption(InputOptionKeys::ARCHIVE_DIR)) {
+        if ($archiveDir = $this->input->getOption(InputOptionKeys::ARCHIVE_DIR)) {
             $instance->setArchiveDir($archiveDir);
         }
 
         // query whether or not the debug mode has been specified as command line
         // option, if yes override the value from the configuration file
-        if ($archiveArtefacts = $input->getOption(InputOptionKeys::ARCHIVE_ARTEFACTS)) {
+        if ($archiveArtefacts = $this->input->getOption(InputOptionKeys::ARCHIVE_ARTEFACTS)) {
             $instance->setArchiveArtefacts($instance->mapBoolean($archiveArtefacts));
         }
 
         // query whether or not a source date format has been specified as command
         // line  option, if yes override the value from the configuration file
-        if ($sourceDateFormat = $input->getOption(InputOptionKeys::SOURCE_DATE_FORMAT)) {
+        if ($sourceDateFormat = $this->input->getOption(InputOptionKeys::SOURCE_DATE_FORMAT)) {
             $instance->setSourceDateFormat($sourceDateFormat);
         }
 
         // query whether or not the debug mode has been specified as command line
         // option, if yes override the value from the configuration file
-        if ($debugMode = $input->getOption(InputOptionKeys::DEBUG_MODE)) {
+        if ($debugMode = $this->input->getOption(InputOptionKeys::DEBUG_MODE)) {
             $instance->setDebugMode($instance->mapBoolean($debugMode));
         }
 
         // query whether or not the log level has been specified as command line
         // option, if yes override the value from the configuration file
-        if ($logLevel = $input->getOption(InputOptionKeys::LOG_LEVEL)) {
+        if ($logLevel = $this->input->getOption(InputOptionKeys::LOG_LEVEL)) {
             $instance->setLogLevel($logLevel);
         }
 
@@ -357,12 +172,12 @@ class ConfigurationLoader
 
         // query whether or not a DB ID has been specified as command line
         // option, if yes override the value from the configuration file
-        if ($useDbId = $input->getOption(InputOptionKeys::USE_DB_ID)) {
+        if ($useDbId = $this->input->getOption(InputOptionKeys::USE_DB_ID)) {
             $instance->setUseDbId($useDbId);
         } else {
             // query whether or not a PDO DSN has been specified as command line
             // option, if yes override the value from the configuration file
-            if ($dsn = $input->getOption(InputOptionKeys::DB_PDO_DSN)) {
+            if ($dsn = $this->input->getOption(InputOptionKeys::DB_PDO_DSN)) {
                 // first REMOVE all other database configurations
                 $instance->clearDatabases();
 
@@ -370,8 +185,8 @@ class ConfigurationLoader
                 $instance->addDatabase(
                     $this->newDatabaseConfiguration(
                         $dsn,
-                        $input->getOption(InputOptionKeys::DB_USERNAME),
-                        $input->getOption(InputOptionKeys::DB_PASSWORD)
+                        $this->input->getOption(InputOptionKeys::DB_USERNAME),
+                        $this->input->getOption(InputOptionKeys::DB_PASSWORD)
                     )
                 );
             }
@@ -396,7 +211,7 @@ class ConfigurationLoader
 
         // query whether or not the debug mode is enabled and log level
         // has NOT been overwritten with a commandline option
-        if ($instance->isDebugMode() && !$input->getOption(InputOptionKeys::LOG_LEVEL)) {
+        if ($instance->isDebugMode() && !$this->input->getOption(InputOptionKeys::LOG_LEVEL)) {
             // set debug log level, if log level has NOT been overwritten on command line
             $instance->setLogLevel(LogLevel::DEBUG);
         }
@@ -409,6 +224,9 @@ class ConfigurationLoader
             )
         );
 
+        // load the extension libraries, if configured
+        $this->libraryLoader->load($instance);
+
         // return the initialized configuration instance
         return $instance;
     }
@@ -420,7 +238,7 @@ class ConfigurationLoader
      *
      * @return boolean TRUE if the directory is a Magento root directory, else FALSE
      */
-    public function isMagentoRootDir($dir)
+    protected function isMagentoRootDir($dir)
     {
         return is_file($this->getMagentoEnv($dir));
     }
@@ -432,7 +250,7 @@ class ConfigurationLoader
      *
      * @return string The path to the Magento file with the environment configuration
      */
-    public function getMagentoEnv($dir)
+    protected function getMagentoEnv($dir)
     {
         return sprintf('%s/app/etc/env.php', $dir);
     }
@@ -446,7 +264,7 @@ class ConfigurationLoader
      * @return array The connection data
      * @throws \Exception Is thrown, if the requested DB connection is not available
      */
-    public function getMagentoDbConnection($dir, $connectionName = 'default')
+    protected function getMagentoDbConnection($dir, $connectionName = 'default')
     {
 
         // load the magento environment
@@ -482,7 +300,7 @@ class ConfigurationLoader
      *
      * @return \TechDivision\Import\Configuration\Jms\Configuration\Database The database configuration instance
      */
-    public function newDatabaseConfiguration($dsn, $username = 'root', $password = null, $default = true, $id = null)
+    protected function newDatabaseConfiguration($dsn, $username = 'root', $password = null, $default = true, $id = null)
     {
 
         // initialize a new database configuration
@@ -517,29 +335,9 @@ class ConfigurationLoader
      *
      * @return string The DSN
      */
-    public function newDsn($host, $dbName, $charset = 'utf8')
+    protected function newDsn($host, $dbName, $charset = 'utf8')
     {
         return sprintf('mysql:host=%s;dbname=%s;charset=%s', $host, $dbName, $charset);
-    }
-
-    /**
-     * Return's the default configuration for the passed Magento Edition and the actual entity type.
-     *
-     * @param string $magentoEdition The Magento Edition to return the configuration for
-     * @param string $entityTypeCode The entity type code to use
-     *
-     * @return string The path to the default configuration
-     */
-    public function getDefaultConfiguration($magentoEdition, $entityTypeCode)
-    {
-        return sprintf(
-            '%s/%s/etc/techdivision-import.json',
-            $this->getVendorDir(),
-            $this->getDefaultConfigurationLibrary(
-                $magentoEdition,
-                $entityTypeCode
-            )
-        );
     }
 
     /**
@@ -550,7 +348,7 @@ class ConfigurationLoader
      * @return array The Magento Edition specific default libraries
      * @throws \Exception Is thrown, if the passed Magento Edition is NOT supported
      */
-    public function getDefaultLibraries($magentoEdition)
+    protected function getDefaultLibraries($magentoEdition)
     {
 
         // query whether or not, default libraries for the passed edition are available
@@ -568,44 +366,6 @@ class ConfigurationLoader
     }
 
     /**
-     * Return's the Magento Edition and entity type's specific default library that contains
-     * the configuration file.
-     *
-     * @param string $magentoEdition The Magento Edition to return the default library for
-     * @param string $entityTypeCode The entity type code to return the default library file for
-     *
-     * @return string The name of the library that contains the default configuration file for the passed Magento Edition and entity type code
-     * @throws \Exception Is thrown, if no default configuration for the passed entity type code is available
-     */
-    public function getDefaultConfigurationLibrary($magentoEdition, $entityTypeCode)
-    {
-
-        // query whether or not, a default configuration file for the passed entity type is available
-        if (isset($this->defaultConfigurations[$edition = strtolower($magentoEdition)])) {
-            if (isset($this->defaultConfigurations[$edition][$entityTypeCode])) {
-                return $this->defaultConfigurations[$edition][$entityTypeCode];
-            }
-
-            // throw an exception, if the passed entity type is not supported
-            throw new \Exception(
-                sprintf(
-                    'Entity Type Code \'%s\' not supported for Magento Edition \'%s\' (MUST be one of catalog_product, catalog_category or eav_attribute)',
-                    $edition,
-                    $entityTypeCode
-                )
-            );
-        }
-
-        // throw an exception, if the passed edition is not supported
-        throw new \Exception(
-            sprintf(
-                'Default configuration for Magento \'%s\' not supported (MUST be one of CE or EE)',
-                $magentoEdition
-            )
-        );
-    }
-
-    /**
      * Return's the entity types specific default import directory.
      *
      * @param string $entityTypeCode The entity type code to return the default import directory for
@@ -613,7 +373,7 @@ class ConfigurationLoader
      * @return string The default default import directory
      * @throws \Exception Is thrown, if no default import directory for the passed entity type code is available
      */
-    public function getDefaultDirectory($entityTypeCode)
+    protected function getDefaultDirectory($entityTypeCode)
     {
 
         // query whether or not, a default configuration file for the passed entity type is available