+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# Version 1.0.0-beta21
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Refactor ConfigurationLoader, add SimpleConfigurationLoader
+* Refactor command classes for usage with SimpleConfigurationLoader
+* Add LibraryLoader class to load external libraries found in configuration
+* Switch to latest library version for better DI support
+
 # Version 1.0.0-beta20
 
 ## Bugfixes